--- a/models/demos/segformer/demo/demo_for_image_classification.py
+++ b/models/demos/segformer/demo/demo_for_image_classification.py
@@ -30,10 +30,20 @@ def run_segformer_classification_demo(
     iterations = iterations // batch_size
     data_loader = get_data_loader(input_loc, batch_size, iterations)
     inputs_mapper, wts_mapper, output_composer = get_mesh_mappers(device)
-
+    config = load_config("configs/segformer_img_classification_config.json")
+    reference_model = SegformerForImageClassificationReference(config)
+    reference_model = load_torch_model(
+        reference_model, f"", module="image_classification", model_location_generator=model_location_generator
+    )
+    parameters = preprocess_model_parameters(
+        initialize_model=lambda: reference_model,
+        custom_preprocessor=create_custom_mesh_preprocessor(wts_mapper),
+        device=None,
+    )
+    parameters = move_to_device(parameters, device)
+    ttnn_model = TtSegformerForImageClassification(config, parameters)
     for iter in range(iterations):
         torch_input_tensor, labels = get_batch(data_loader, resolution)
-        # torch_input_tensor_permuted = torch.permute(torch_input_tensor, (0, 2, 3, 1))
         ttnn_input_tensor = ttnn.from_torch(
             torch_input_tensor,
             dtype=ttnn.bfloat16,
@@ -42,19 +52,7 @@ def run_segformer_classification_demo(
             layout=ttnn.ROW_MAJOR_LAYOUT,
             mesh_mapper=inputs_mapper,
         )
-        config = load_config("configs/segformer_img_classification_config.json")
-        reference_model = SegformerForImageClassificationReference(config)
-        reference_model = load_torch_model(
-            reference_model, f"", module="image_classification", model_location_generator=model_location_generator
-        )
         torch_output = reference_model(torch_input_tensor)
-        parameters = preprocess_model_parameters(
-            initialize_model=lambda: reference_model,
-            custom_preprocessor=create_custom_mesh_preprocessor(wts_mapper),
-            device=None,
-        )
-        parameters = move_to_device(parameters, device)
-        ttnn_model = TtSegformerForImageClassification(config, parameters)
         ttnn_output = ttnn_model(
             ttnn_input_tensor,
             output_attentions=None,