--- a/models/demos/segformer/runner/performant_runner_infra.py
+++ b/models/demos/segformer/runner/performant_runner_infra.py
@@ -3,8 +3,9 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
+import time
+
 import torch
-import torch.nn.functional as F
 from loguru import logger
 from ttnn.model_preprocessing import ParameterDict, ParameterList, preprocess_model_parameters
 
@@ -117,7 +118,8 @@ class SegformerTestInfra:
             parameters=self.parameters,
         )
 
-    def setup_l1_sharded_input(self, device, torch_input_tensor=None):
+    def setup_l1_sharded_input(self, device, torch_input_tensor=None, min_channels=16):
+        assignment_time = time.time()
         if is_wormhole_b0():
             core_grid = ttnn.CoreGrid(y=8, x=8)
         else:
@@ -126,23 +128,47 @@ class SegformerTestInfra:
         # torch tensor
         torch_input_tensor = self.torch_input if torch_input_tensor is None else torch_input_tensor
         n, c, h, w = torch_input_tensor.shape
-        n = n // self.num_devices
-        # sharded mem config for fold input
-        num_cores = core_grid.x * core_grid.y
-        shard_h = (n * w * h + num_cores - 1) // num_cores
-        grid_size = core_grid
-        grid_coord = ttnn.CoreCoord(grid_size.x - 1, grid_size.y - 1)
-        shard_grid = ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), grid_coord)})
-        shard_spec = ttnn.ShardSpec(shard_grid, (shard_h, 16), ttnn.ShardOrientation.ROW_MAJOR)
-        input_mem_config = ttnn.MemoryConfig(
-            ttnn.types.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.types.BufferType.L1, shard_spec
+        if c < min_channels:
+            c = min_channels
+        elif c % min_channels != 0:
+            c = ((c // min_channels) + 1) * min_channels
+
+        assert n % self.num_devices == 0, f"n isn't evenly divided by the available number of devices"
+        n = n // self.num_devices if n // self.num_devices != 0 else n
+        # n = n // self.num_devices
+        # # sharded mem config for fold input
+        # num_cores = core_grid.x * core_grid.y
+        # shard_h = (n * w * h + num_cores - 1) // num_cores
+        # grid_size = core_grid
+        # grid_coord = ttnn.CoreCoord(grid_size.x - 1, grid_size.y - 1)
+        # shard_grid = ttnn.CoreRangeSet({ttnn.CoreRange(ttnn.CoreCoord(0, 0), grid_coord)})
+        # shard_spec = ttnn.ShardSpec(shard_grid, (shard_h, 16), ttnn.ShardOrientation.ROW_MAJOR)
+        print("assignment of ops time:", time.time() - assignment_time, "seconds")
+        config_create_time = time.time()
+        # input_mem_config = ttnn.MemoryConfig(
+        #     ttnn.types.TensorMemoryLayout.HEIGHT_SHARDED, ttnn.types.BufferType.L1, shard_spec
+        # )
+        input_mem_config = ttnn.create_sharded_memory_config(
+            [n, c, h, w],
+            ttnn.CoreGrid(x=8, y=8),
+            ttnn.ShardStrategy.HEIGHT,
         )
-        torch_input_tensor = torch_input_tensor.permute(0, 2, 3, 1)
-        torch_input_tensor = F.pad(torch_input_tensor, (0, 13))
+        print("config creatiuontime:", time.time() - config_create_time, "seconds")
+        # permute_time = time.time()
+        # torch_input_tensor = torch_input_tensor.permute(0, 2, 3, 1)
+        # print(" permute_time time:", time.time() - permute_time, "seconds")
+        # pad_time = time.time()
+        # torch_input_tensor = F.pad(torch_input_tensor, (0, 13))
+        # print(" pad_time time:", time.time() - pad_time, "seconds")
+        from_host_time_prep = time.time()
         input_tensor = [torch_input_tensor[i].unsqueeze(0) for i in range(torch_input_tensor.shape[0])]
+        print(" from_host_time_prep time:", time.time() - from_host_time_prep, "seconds")
+        # tt_inputs_host = ttnn.from_torch(torch_input_tensor, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT,mesh_mapper=self.input_mesh_mapper)
+        from_host_time = time.time()
         tt_inputs_host = ttnn.from_host_shards(
             [ttnn.from_torch(t, dtype=ttnn.bfloat16, layout=ttnn.ROW_MAJOR_LAYOUT) for t in input_tensor], device.shape
         )
+        print("from_host_timetime:", time.time() - from_host_time, "seconds")
         return tt_inputs_host, input_mem_config
 
     def setup_dram_sharded_input(self, device, torch_input_tensor=None):
@@ -154,7 +180,7 @@ class SegformerTestInfra:
             ),
             [
                 divup(tt_inputs_host.volume() // tt_inputs_host.shape[-1], (dram_grid_size.x * dram_grid_size.y)),
-                16,
+                tt_inputs_host.shape[-1],
             ],
             ttnn.ShardOrientation.ROW_MAJOR,
         )