--- a/models/demos/segformer/tests/pcc/test_segformer_model.py
+++ b/models/demos/segformer/tests/pcc/test_segformer_model.py
@@ -89,11 +89,11 @@ def test_segformer_model(batch_size, num_channels, height, width, device, model_
     # # adjust padding if necessary
     # if num_channels < CONV2D_MIN_CHANNEL_SIZE:
     #     ttnn_input_tensor = ttnn.pad(
-    #         ttnn_input_tensor, [batch_size, height, width, CONV2D_MIN_CHANNEL_SIZE], [0, 0, 0, 0], 0
+    #         ttnn_input_tensor, [batch_size,CONV2D_MIN_CHANNEL_SIZE, height, width], [0, 0, 0, 0], 0
     #     )
     # elif num_channels > CONV2D_MIN_CHANNEL_SIZE and num_channels % 32 != 0:
     #     ttnn_input_tensor = ttnn.pad(
-    #         ttnn_input_tensor, [batch_size, height, width, (num_channels + 31) // 32 * 32], [0, 0, 0, 0], 0
+    #         ttnn_input_tensor, [batch_size,(num_channels + 31) // 32 * 32, height, width], [0, 0, 0, 0], 0
     #     )
 
     ttnn_input_tensor = ttnn.to_device(ttnn_input_tensor, device, memory_config=ttnn.L1_MEMORY_CONFIG)