… models/demos/segformer/README.md
… models/demos/segformer/tests/perf/test_e2e_performant.py
… models/demos/segformer/tests/perf/test_perf_segformer.py