### Ticket
Link to Github Issue

### Problem description
DP e2e perf for Segformer on Main is lower than expected
Current Perf Details: 
- (Single-card) - 98 FPS
- DP(N300) - 145 FPS

### What's changed
- torch pad op as part of pre-processing (setup_l1_sharded_input) is responsible for this discrepancy.
- Moved pad/permute inside the model.
- Update Perf Details: (Single-card) - 103 FPS, DP(N300) - 202 FPS

### Checklist
- [ ] [APC]()
- [ ] [Device Perf]()
- [ ] [Model Perf]()
- [ ] [PCC tests]()
- [ ] [Demo tests]()